+++ github_macros/__init__.py
@@ -1,2 +1,2 @@
 from __future__ import print_function
-__version__ = '2.0.0'
+__version__ = '2.1.0'